+++ src/pages/product/ProductViewPage.tsx
@@ -3,7 +3,6 @@ import { ReactComponent as Star } from '@/assets/svgs/star.svg';
 import product1 from '@/assets/images/product1.jpeg';
 import profile from '@/assets/images/profile.png';
 import { ProductStatusChip } from '@/components/Chips';
-import { NavigateButton, ProductRelatedButton } from './components/button';
 import { useNavigate } from 'react-router-dom';
 import { ReviewCard } from './components/ReviewCard';
 import {
@@ -13,6 +12,8 @@ import {
   CarouselNext,
   CarouselPrevious,
 } from '@/components/ui/carousel';
+import { ProductRelatedButton } from './components/button/ProductRelatedButton';
+import { NavigateButton } from './components/button/NavigateButton';
 
 export const ProductViewPage = () => {
   const navigate = useNavigate();
@@ -24,6 +25,10 @@ export const ProductViewPage = () => {
   const handleNavigateEdit = () => {
     navigate('/product/1/edit');
   };
+
+  const tags = [
+    '#가전제품', '#노트북', '#애플', '#맥북', '#실버'
+  ];
 
   return (
     <div className="min-h-screen flex w-screen pb-[133px]">
@@ -52,9 +57,21 @@ export const ProductViewPage = () => {
             </div>
           </div>
           <div className="flex w-full justify-between">
-            <span className="text-xlarge28 font-semibold text-neutral-0">
-              맥북 프로 실버 완전 싸게 대여하세요~
-            </span>
+            <div className='flex flex-col'>
+              <span className="text-xlarge28 font-semibold text-neutral-0">
+                맥북 프로 실버 완전 싸게 대여하세요~
+              </span>
+              <div className="flex gap-2 mt-2">
+                {tags.map((tag, index) => (
+                  <span
+                    key={index}
+                    className="bg-neutral-100 px-2 py-1 rounded-md text-neutral-0 text-xsmall14"
+                  >
+                    {tag}
+                  </span>
+                ))}
+              </div>
+            </div>
             <div className="flex items-center space-x-2">
               <img
                 src={profile}
